+++ import_export/admin.py
@@ -2,12 +2,15 @@ from __future__ import with_statement
 
 import tempfile
 from datetime import datetime
+import os.path
 
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 from django.conf.urls import patterns, url
 from django.template.response import TemplateResponse
 from django.contrib import messages
+from django.contrib.admin.models import LogEntry, ADDITION, CHANGE, DELETION
+from django.contrib.contenttypes.models import ContentType
 from django.http import HttpResponseRedirect, HttpResponse
 from django.core.urlresolvers import reverse
 
@@ -20,6 +23,12 @@ from .resources import (
     modelresource_factory,
 )
 from .formats import base_formats
+from .results import RowResult
+
+try:
+    from django.utils.encoding import force_text
+except ImportError:
+    from django.utils.encoding import force_unicode as force_text
 
 
 #: import / export formats
@@ -70,6 +79,12 @@ class ImportMixin(object):
         else:
             return self.resource_class
 
+    def get_import_resource_class(self):
+        """
+        Returns ResourceClass to use for import.
+        """
+        return self.get_resource_class()
+
     def get_import_formats(self):
         """
         Returns available import formats.
@@ -82,7 +97,7 @@ class ImportMixin(object):
         wishes to import)
         '''
         opts = self.model._meta
-        resource = self.get_resource_class()()
+        resource = self.get_import_resource_class()()
 
         confirm_form = ConfirmImportForm(request.POST)
         if confirm_form.is_valid():
@@ -90,15 +105,35 @@ class ImportMixin(object):
             input_format = import_formats[
                 int(confirm_form.cleaned_data['input_format'])
             ]()
-            import_file = open(confirm_form.cleaned_data['import_file_name'],
-                               input_format.get_read_mode())
+            import_file_name = os.path.join(
+                tempfile.gettempdir(),
+                confirm_form.cleaned_data['import_file_name']
+            )
+            import_file = open(import_file_name, input_format.get_read_mode())
             data = import_file.read()
             if not input_format.is_binary() and self.from_encoding:
-                data = unicode(data, self.from_encoding).encode('utf-8')
+                data = force_text(data, self.from_encoding)
             dataset = input_format.create_dataset(data)
 
-            resource.import_data(dataset, dry_run=False,
+            result = resource.import_data(dataset, dry_run=False,
                                  raise_errors=True)
+
+            # Add imported objects to LogEntry
+            logentry_map = {
+                RowResult.IMPORT_TYPE_NEW: ADDITION,
+                RowResult.IMPORT_TYPE_UPDATE: CHANGE,
+                RowResult.IMPORT_TYPE_DELETE: DELETION,
+            }
+            content_type_id=ContentType.objects.get_for_model(self.model).pk
+            for row in result:
+                LogEntry.objects.log_action(
+                    user_id=request.user.pk,
+                    content_type_id=content_type_id,
+                    object_id=row.object_id,
+                    object_repr=row.object_repr,
+                    action_flag=logentry_map[row.import_type],
+                    change_message="%s through import_export" % row.import_type,
+                )
 
             success_message = _('Import finished')
             messages.success(request, success_message)
@@ -116,7 +151,7 @@ class ImportMixin(object):
         uploaded file to a local temp file that will be used by
         'process_import' for the actual import.
         '''
-        resource = self.get_resource_class()()
+        resource = self.get_import_resource_class()()
 
         context = {}
 
@@ -142,7 +177,7 @@ class ImportMixin(object):
                 # warning, big files may exceed memory
                 data = uploaded_import_file.read()
                 if not input_format.is_binary() and self.from_encoding:
-                    data = unicode(data, self.from_encoding).encode('utf-8')
+                    data = force_text(data, self.from_encoding)
                 dataset = input_format.create_dataset(data)
                 result = resource.import_data(dataset, dry_run=True,
                                               raise_errors=False)
@@ -151,7 +186,7 @@ class ImportMixin(object):
 
             if not result.has_errors():
                 context['confirm_form'] = ConfirmImportForm(initial={
-                    'import_file_name': uploaded_file.name,
+                    'import_file_name': os.path.basename(uploaded_file.name),
                     'input_format': form.cleaned_data['input_format'],
                 })
 
@@ -195,6 +230,12 @@ class ExportMixin(object):
         else:
             return self.resource_class
 
+    def get_export_resource_class(self):
+        """
+        Returns ResourceClass to use for export.
+        """
+        return self.get_resource_class()
+
     def get_export_formats(self):
         """
         Returns available import formats.
@@ -236,7 +277,7 @@ class ExportMixin(object):
                 int(form.cleaned_data['file_format'])
             ]()
 
-            resource_class = self.get_resource_class()
+            resource_class = self.get_export_resource_class()
             queryset = self.get_export_queryset(request)
             data = resource_class().export(queryset)
             response = HttpResponse(