+++ import_export/results.py
@@ -1,3 +1,6 @@
+from __future__ import unicode_literals
+
+
 class Error(object):
 
     def __init__(self, error, traceback=None):
@@ -30,3 +33,6 @@ class Result(object):
 
     def has_errors(self):
         return bool(self.base_errors or self.row_errors())
+
+    def __iter__(self):
+        return iter(self.rows)